requirements.txt
fastapi
pydantic
google-generativeai